+++ app/models/survey/AnswerFormat.java
@@ -1,6 +1,6 @@
 /*******************************************************************************
  *        File: AnswerFormat.java
- *    Revision: 2
+ *    Revision: 3
  * Description: 
  *      Author: Morteza Ansarinia <ansarinia@me.com>
  *  Created on: July 3, 2015
@@ -9,10 +9,22 @@
  *******************************************************************************/
 package models.survey;
 
-//TODO make it an enitity to apply more complex answer formats.
 public enum AnswerFormat {
   STATEMENT,
   TEXT,
   MULTIPLE_CHOICE,
-  SCALE
+  SCALE;
+  
+  public static AnswerFormat parse(String type) {
+    if (type.equalsIgnoreCase("statement") || type.equalsIgnoreCase("text")) {
+      return STATEMENT;
+    }else if (type.equalsIgnoreCase("text_question")) {
+      return TEXT;
+    }else if (type.equalsIgnoreCase("scale_question")) {
+      return SCALE;
+    }else if (type.equalsIgnoreCase("multiple_choice_question")) {
+      return MULTIPLE_CHOICE;
+    }
+    return null;
+  }
 }